backend/requirements.txt
fastapi
uvicorn
gtts
mysql-connector-python
numpy